requirements.txt
Pillow>=10,<13
pybind11>=2.11,<3
